Fixes #12197: More clean-up on blank template project
