+++ blank-template/src/main/resources/template/BLANKManagement.html
